added an overview of the data to expect from the main scraper along with an example file
