+++ README.md
@@ -1,5 +1,5 @@
 # hacktx
-instructions to run:
+instructions to run the scrape:
   - clone repo to local folder
   - run npm install
   - make an index.js file
@@ -7,3 +7,5 @@ instructions to run:
   - make an anonymous function and have it be async
   - initialize a variable and assign it to the scrapeAllProjectsFromUser function and pass a parameter of a user's username as a string
   - console.log the variable
+
+  - also see and run 'example.js' by using node example.js in the terminal
