+++ example.js
@@ -1,0 +1,6 @@
+const scrapeUser = require("./webScraper/scrapeUser");
+
+(async() => {
+    const data = await scrapeUser("12trinhle")
+    console.log(data.projects)
+})();
